refactor: 일기 작성 페이지를 리팩토링하였습니다.
